+++ packages/requisitions/src/RequestRequisition/DetailView/AppBarButtons/AppBarButtons.tsx
@@ -15,7 +15,7 @@ import {
   ReportSelector,
   useReport,
 } from '@openmsupply-client/system';
-import { useIsRequestDisabled, useRequest } from '../../api';
+import { useRequest } from '../../api';
 import { UseSuggestedQuantityButton } from './UseSuggestedQuantityButton';
 import { AddFromMasterListButton } from './AddFromMasterListButton';
 
@@ -27,10 +27,10 @@ interface AppBarButtonProps {
 export const AppBarButtonsComponent: FC<AppBarButtonProps> = ({
   onAddItem,
 }) => {
-  const isDisabled = useIsRequestDisabled();
+  const isDisabled = useRequest.utils.isDisabled();
   const { OpenButton } = useDetailPanel();
   const t = useTranslation('distribution');
-  const { data } = useRequest();
+  const { data } = useRequest.document.get();
   const { print, isPrinting } = useReport.utils.print();
 
   const printReport = (report: ReportRowFragment) => {
