+++ packages/requisitions/src/RequestRequisition/DetailView/RequestLineEdit/ItemCharts/StockEvolution.tsx
@@ -14,7 +14,7 @@ import {
 } from '@common/components';
 import { Box, useTheme, useTranslation } from '@openmsupply-client/common';
 import { useFormatDateTime } from '@common/intl';
-import { useRequestLineChartData } from '../../../api/hooks';
+import { useRequest } from '../../../api/hooks';
 
 export interface StockEvolutionProps {
   id: string;
@@ -24,7 +24,7 @@ export const StockEvolution: React.FC<StockEvolutionProps> = ({ id }) => {
   const t = useTranslation('replenishment');
   const theme = useTheme();
   const { dayMonthShort } = useFormatDateTime();
-  const { data, isLoading } = useRequestLineChartData(id);
+  const { data, isLoading } = useRequest.line.chartData(id);
 
   const dateFormatter = (date: string) => dayMonthShort(date);
   const tooltipFormatter = (value: number, name: string) => {
