+++ packages/requisitions/src/RequestRequisition/api/hooks/document/index.ts
@@ -1,0 +1,19 @@
+import { useDeleteRequests } from './useDeleteRequests';
+import { useDeleteSelectedRequisitions } from './useDeleteSelectedRequisitions';
+import { useInsertRequest } from './useInsertRequest';
+import { useRequest } from './useRequest';
+import { useRequestFields } from './useRequestFields';
+import { useRequests } from './useRequests';
+import { useRequestsAll } from './useRequestsAll';
+import { useUpdateRequest } from './useUpdateRequest';
+
+export const Document = {
+  useDeleteRequests,
+  useDeleteSelectedRequisitions,
+  useInsertRequest,
+  useRequest,
+  useRequestFields,
+  useRequests,
+  useRequestsAll,
+  useUpdateRequest,
+};
