+++ packages/requisitions/src/RequestRequisition/api/hooks/document/useDeleteRequests.ts
@@ -1,0 +1,10 @@
+import { useQueryClient, useMutation } from '@openmsupply-client/common';
+import { useRequestApi } from '../utils/useRequestApi';
+
+export const useDeleteRequests = () => {
+  const queryClient = useQueryClient();
+  const api = useRequestApi();
+  return useMutation(api.deleteRequests, {
+    onSuccess: () => queryClient.invalidateQueries(api.keys.base()),
+  });
+};
