+++ packages/requisitions/src/RequestRequisition/api/hooks/document/useInsertRequest.ts
@@ -1,0 +1,18 @@
+import {
+  useQueryClient,
+  useNavigate,
+  useMutation,
+} from '@openmsupply-client/common';
+import { useRequestApi } from '../utils/useRequestApi';
+
+export const useInsertRequest = () => {
+  const queryClient = useQueryClient();
+  const navigate = useNavigate();
+  const api = useRequestApi();
+  return useMutation(api.insert, {
+    onSuccess: ({ requisitionNumber }) => {
+      navigate(String(requisitionNumber));
+      queryClient.invalidateQueries(api.keys.base());
+    },
+  });
+};
