+++ packages/requisitions/src/RequestRequisition/api/hooks/document/useUpdateRequest.ts
@@ -1,0 +1,10 @@
+import { useQueryClient, useMutation } from '@openmsupply-client/common';
+import { useRequestApi } from '../utils/useRequestApi';
+
+export const useUpdateRequest = () => {
+  const queryClient = useQueryClient();
+  const api = useRequestApi();
+  return useMutation(api.update, {
+    onSuccess: () => queryClient.invalidateQueries(api.keys.base()),
+  });
+};
