+++ packages/requisitions/src/RequestRequisition/api/hooks/line/index.ts
@@ -1,0 +1,11 @@
+import { useDeleteRequestLines } from './useDeleteRequestLines';
+import { useRequestLineChartData } from './useRequestLineChartData';
+import { useRequestLines } from './useRequestLines';
+import { useSaveRequestLines } from './useSaveRequestLines';
+
+export const Lines = {
+  useDeleteRequestLines,
+  useRequestLineChartData,
+  useRequestLines,
+  useSaveRequestLines,
+};
