+++ packages/requisitions/src/RequestRequisition/api/hooks/utils/index.ts
@@ -1,0 +1,11 @@
+import { useRequestApi } from './useRequestApi';
+import { useAddFromMasterList } from './useAddFromMasterList';
+import { useIsRequestDisabled } from './useIsRequestDisabled';
+import { useSuggestedQuantity } from './useSuggestedQuantity';
+
+export const Utils = {
+  useRequestApi,
+  useAddFromMasterList,
+  useIsRequestDisabled,
+  useSuggestedQuantity,
+};
