+++ packages/requisitions/src/RequestRequisition/api/hooks/utils/useIsRequestDisabled.ts
@@ -1,0 +1,8 @@
+import { useRequest } from '../document/useRequest';
+import { isRequestDisabled } from './../../../../utils';
+
+export const useIsRequestDisabled = (): boolean => {
+  const { data } = useRequest();
+  if (!data) return true;
+  return isRequestDisabled(data);
+};
